requirements.txt
# pyttsx3==2.98
# pywhatkit==5.4
typeguard==2.12.1
sounddevice==0.5.1
SpeechRecognition==3.14.1
wikipedia==1.4.0
wave
pyserial==3.3
pynput
pyyaml
jinja2
PyAudio==0.2.14
keyboard
pydub
ffmpeg
numpy